requirements.txt
labml==0.4.168
labml_helpers==0.4.89
labml_nn==0.4.137
matplotlib==3.8.2
numpy==2.1.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
pandas==2.2.3
Pillow==9.5.0
Pillow==11.0.0
scikit_learn==1.5.2
seaborn==0.13.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.66.1
wavemix==0.2.4
